+++ packages/node/src/index.ts
@@ -5,4 +5,4 @@ export { ChatProviderIdEnum, PushProviderIdEnum };
 export * from './lib/novu';
 export * from './lib/subscribers/subscriber.interface';
 export * from './lib/events/events.interface';
-export * from './lib/notificationGroups/notificationGroups.interface';
+export * from './lib/notification-groups/notification-groups.interface';
