+++ packages/node/src/lib/notification-groups/notification-groups.interface.ts
